+++ .metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.jdt.ui.prefs
@@ -1,10 +1,100 @@
-content_assist_proposals_background=255,255,255
-content_assist_proposals_foreground=0,0,0
+content_assist_completion_replacement_background=200,200,0
+content_assist_completion_replacement_foreground=200,0,0
+content_assist_number_of_computers=21
+content_assist_parameters_background=52,57,61
+content_assist_parameters_foreground=238,238,238
+content_assist_proposals_background=52,57,61
+content_assist_proposals_foreground=238,238,238
 eclipse.preferences.version=1
+java_bracket=249,250,244
+java_comment_task_tag=154,140,124
+java_default=217,232,247
+java_doc_default=128,128,128
+java_doc_keyword=154,140,124
+java_doc_link=169,156,140
+java_doc_tag=30,120,155
+java_keyword=204,108,29
+java_keyword_bold=false
+java_keyword_return=204,108,29
+java_keyword_return_bold=false
+java_multi_line_comment=128,128,128
+java_operator=230,230,250
+java_single_line_comment=128,128,128
+java_string=23,198,163
+matchingBracketsColor=249,250,244
 org.eclipse.jdt.internal.ui.navigator.layout=2
 org.eclipse.jdt.internal.ui.navigator.librariesnode=true
 org.eclipse.jdt.ui.formatterprofiles.version=23
+overriddenByCSS=,content_assist_completion_replacement_background,content_assist_completion_replacement_foreground,content_assist_parameters_background,content_assist_parameters_foreground,content_assist_proposals_background,content_assist_proposals_foreground,java_bracket,java_comment_task_tag,java_default,java_doc_default,java_doc_keyword,java_doc_link,java_doc_tag,java_keyword,java_keyword_bold,java_keyword_return,java_keyword_return_bold,java_multi_line_comment,java_operator,java_single_line_comment,java_string,matchingBracketsColor,pf_coloring_argument,pf_coloring_assignment,pf_coloring_comment,pf_coloring_key,pf_coloring_value,semanticHighlighting.abstractClass.color,semanticHighlighting.abstractClass.enabled,semanticHighlighting.abstractMethodInvocation.color,semanticHighlighting.abstractMethodInvocation.enabled,semanticHighlighting.annotation.color,semanticHighlighting.annotation.enabled,semanticHighlighting.annotation.italic,semanticHighlighting.annotationElementReference.color,semanticHighlighting.annotationElementReference.enabled,semanticHighlighting.class.color,semanticHighlighting.class.enabled,semanticHighlighting.deprecatedMember.color,semanticHighlighting.deprecatedMember.enabled,semanticHighlighting.deprecatedMember.underline,semanticHighlighting.deprecatedMember.strikethrough,semanticHighlighting.enum.color,semanticHighlighting.enum.enabled,semanticHighlighting.enum.italic,semanticHighlighting.field.color,semanticHighlighting.field.enabled,semanticHighlighting.inheritedField.color,semanticHighlighting.inheritedMethodInvocation.color,semanticHighlighting.inheritedMethodInvocation.enabled,semanticHighlighting.interface.color,semanticHighlighting.interface.enabled,semanticHighlighting.localVariable.color,semanticHighlighting.localVariable.enabled,semanticHighlighting.localVariableDeclaration.color,semanticHighlighting.localVariableDeclaration.enabled,semanticHighlighting.localVariableDeclaration.bold,semanticHighlighting.method.color,semanticHighlighting.method.enabled,semanticHighlighting.methodDeclarationName.color,semanticHighlighting.methodDeclarationName.enabled,semanticHighlighting.methodDeclarationName.bold,semanticHighlighting.number.color,semanticHighlighting.number.enabled,semanticHighlighting.parameterVariable.color,semanticHighlighting.parameterVariable.enabled,semanticHighlighting.staticField.color,semanticHighlighting.staticField.enabled,semanticHighlighting.staticFinalField.color,semanticHighlighting.staticFinalField.enabled,semanticHighlighting.staticMethodInvocation.color,semanticHighlighting.staticMethodInvocation.enabled,semanticHighlighting.typeArgument.color,semanticHighlighting.typeArgument.enabled,semanticHighlighting.typeParameter.color,semanticHighlighting.typeParameter.enabled,semanticHighlighting.typeParameter.bold,semanticHighlighting.restrictedKeywords.color,semanticHighlighting.restrictedKeywords.bold,sourceHoverBackgroundColor,
+pf_coloring_argument=221,40,103
+pf_coloring_assignment=217,232,247
+pf_coloring_comment=128,128,128
+pf_coloring_key=217,232,247
+pf_coloring_value=23,198,163
+semanticHighlighting.abstractClass.color=62,171,230
+semanticHighlighting.abstractClass.enabled=true
+semanticHighlighting.abstractMethodInvocation.color=128,246,167
+semanticHighlighting.abstractMethodInvocation.enabled=true
+semanticHighlighting.annotation.color=160,160,160
+semanticHighlighting.annotation.enabled=true
+semanticHighlighting.annotation.italic=true
+semanticHighlighting.annotationElementReference.color=235,75,100
+semanticHighlighting.annotationElementReference.enabled=true
+semanticHighlighting.class.color=18,144,195
+semanticHighlighting.class.enabled=true
+semanticHighlighting.deprecatedMember.color=128,128,128
+semanticHighlighting.deprecatedMember.enabled=true
+semanticHighlighting.deprecatedMember.strikethrough=true
+semanticHighlighting.deprecatedMember.underline=false
+semanticHighlighting.enum.color=204,129,186
+semanticHighlighting.enum.enabled=true
+semanticHighlighting.enum.italic=true
+semanticHighlighting.field.color=102,225,248
+semanticHighlighting.field.enabled=true
+semanticHighlighting.inheritedField.color=143,143,191
+semanticHighlighting.inheritedMethodInvocation.color=205,246,104
+semanticHighlighting.inheritedMethodInvocation.enabled=true
+semanticHighlighting.interface.color=128,242,246
+semanticHighlighting.interface.enabled=true
+semanticHighlighting.localVariable.color=243,236,121
+semanticHighlighting.localVariable.enabled=true
+semanticHighlighting.localVariableDeclaration.bold=false
+semanticHighlighting.localVariableDeclaration.color=242,242,0
+semanticHighlighting.localVariableDeclaration.enabled=true
+semanticHighlighting.method.color=167,236,33
+semanticHighlighting.method.enabled=true
+semanticHighlighting.methodDeclarationName.bold=false
+semanticHighlighting.methodDeclarationName.color=30,181,64
+semanticHighlighting.methodDeclarationName.enabled=true
+semanticHighlighting.number.color=104,151,187
+semanticHighlighting.number.enabled=true
+semanticHighlighting.parameterVariable.color=121,171,255
+semanticHighlighting.parameterVariable.enabled=true
+semanticHighlighting.restrictedKeywords.bold=false
+semanticHighlighting.restrictedKeywords.color=204,108,29
+semanticHighlighting.staticField.color=141,218,248
+semanticHighlighting.staticField.enabled=true
+semanticHighlighting.staticFinalField.color=141,218,248
+semanticHighlighting.staticFinalField.enabled=true
+semanticHighlighting.staticMethodInvocation.color=150,236,63
+semanticHighlighting.staticMethodInvocation.enabled=true
+semanticHighlighting.typeArgument.color=177,102,218
+semanticHighlighting.typeArgument.enabled=true
+semanticHighlighting.typeParameter.bold=false
+semanticHighlighting.typeParameter.color=191,164,164
+semanticHighlighting.typeParameter.enabled=true
+sourceHoverBackgroundColor=68,68,68
+spelling_ignore_ampersand_in_properties=true
+spelling_ignore_digits=true
+spelling_ignore_java_strings=true
+spelling_ignore_mixed=true
+spelling_ignore_non_letters=true
+spelling_ignore_sentence=true
+spelling_ignore_single_letters=true
+spelling_ignore_upper=true
+spelling_ignore_urls=true
 spelling_locale_initialized=true
+spelling_user_dictionary_encoding=
 typefilter_migrated_2=true
 useAnnotationsPrefPage=true
 useQuickDiffPrefPage=true
