+++ .metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.ui.editors.prefs
@@ -1,2 +1,42 @@
+AbstractTextEditor.Color.Background=47,47,47
+AbstractTextEditor.Color.Background.SystemDefault=false
+AbstractTextEditor.Color.FindScope=30,120,155
+AbstractTextEditor.Color.Foreground=204,204,204
+AbstractTextEditor.Color.Foreground.SystemDefault=false
+AbstractTextEditor.Color.SelectionBackground=33,66,131
+AbstractTextEditor.Color.SelectionBackground.SystemDefault=false
+AbstractTextEditor.Color.SelectionForeground=147,161,161
+AbstractTextEditor.Color.SelectionForeground.SystemDefault=false
+LSP4EReadOccurrenceIndicationColor=27,98,145
+LSP4ETextOccurrenceIndicationColor=27,98,145
+LSP4EWriteOccurrenceIndicationColor=27,98,145
+TM4EMatchingPairColor=27,98,145
+TextOccurrenceIndicationColor=27, 98, 145
+asOccurencesIndicationColor=72,72,72
+breakpointIndicationColor=51,119,193
+currentIPColor=90,90,90
+currentLineColor=55,55,55
+deletionIndicationColor=224,226,228
 eclipse.preferences.version=1
+filteredSearchResultIndicationColor=27,98,145
+fullcoverage_color=6,74,6
+hyperlinkColor=102,175,249
+hyperlinkColor.SystemDefault=false
+infoIndicationColor=86,194,170
+lineNumberColor=119,145,154
+linked.slave.color=66,156,255
+matchingTagIndicationColor=72,72,72
+nocoverage_color=98,5,5
+occurrenceIndicationColor=27,98,145
+overriddenByCSS=,fullcoverage_color,partialcoverage_color,nocoverage_color,LSP4EReadOccurrenceIndicationColor,LSP4EWriteOccurrenceIndicationColor,LSP4ETextOccurrenceIndicationColor,TM4EMatchingPairColor,TextOccurrenceIndicationColor,AbstractTextEditor.Color.Background.SystemDefault,AbstractTextEditor.Color.SelectionForeground.SystemDefault,AbstractTextEditor.Color.SelectionBackground.SystemDefault,AbstractTextEditor.Color.Background,AbstractTextEditor.Color.Foreground.SystemDefault,AbstractTextEditor.Color.SelectionBackground,AbstractTextEditor.Color.SelectionForeground,AbstractTextEditor.Color.Foreground,AbstractTextEditor.Color.FindScope,asOccurencesIndicationColor,breakpointIndicationColor,currentIPColor,currentLineColor,deletionIndicationColor,filteredSearchResultIndicationColor,hyperlinkColor,hyperlinkColor.SystemDefault,infoIndicationColor,lineNumberColor,linked.slave.color,matchingTagIndicationColor,occurrenceIndicationColor,overrideIndicatorColor,printMarginColor,searchResultHighlighting,searchResultIndication,searchResultIndicationColor,searchResultTextStyle,secondaryIPColor,spellingIndicationColor,writeOccurrenceIndicationColor,
+overrideIndicatorColor=78,120,117
+partialcoverage_color=98,98,0
+printMarginColor=81,86,88
+searchResultHighlighting=false
+searchResultIndication=true
+searchResultIndicationColor=94,94,94
+searchResultTextStyle=BOX
+secondaryIPColor=90,90,90
 spellingEnabled=false
+spellingIndicationColor=253,170,211
+writeOccurrenceIndicationColor=27,98,145
