+++ .metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.ui.ide.prefs
@@ -1,4 +1,4 @@
 eclipse.preferences.version=1
-platformState=1706674663157
+platformState=1706674663181
 quickStart=false
 tipsAndTricks=true
