+++ .metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.ui.workbench.prefs
@@ -1,10 +1,73 @@
 //org.eclipse.ui.commands/state/org.eclipse.ui.navigator.resources.nested.changeProjectPresentation/org.eclipse.ui.commands.radioState=false
 //org.eclipse.ui.commands/state/org.eclipse.wst.xml.views.XPathView.processor.xpathprocessor/org.eclipse.ui.commands.radioState=xpath10
+ACTIVE_HYPERLINK_COLOR=138,201,242
+CONFLICTING_COLOR=240,15,66
+CONTENT_ASSIST_BACKGROUND_COLOR=52,57,61
+CONTENT_ASSIST_FOREGROUND_COLOR=238,238,238
+EDITION_COLOR=238,238,238
+ERROR_COLOR=247,68,117
+HYPERLINK_COLOR=111,197,238
+INCOMING_COLOR=31,179,235
+OUTGOING_COLOR=238,238,238
 PLUGINS_NOT_ACTIVATED_ON_STARTUP=;org.eclipse.m2e.discovery;
+RESOLVED_COLOR=108,210,17
 eclipse.preferences.version=1
-org.eclipse.ui.workbench.ACTIVE_NOFOCUS_TAB_BG_END=255,255,255
-org.eclipse.ui.workbench.ACTIVE_NOFOCUS_TAB_BG_START=255,255,255
-org.eclipse.ui.workbench.ACTIVE_NOFOCUS_TAB_TEXT_COLOR=16,16,16
-org.eclipse.ui.workbench.ACTIVE_TAB_BG_END=255,255,255
-org.eclipse.ui.workbench.ACTIVE_TAB_BG_START=255,255,255
-org.eclipse.ui.workbench.INACTIVE_TAB_BG_START=242,242,242
+org.eclipse.egit.ui.CommitMessageCommentColor=128,128,128
+org.eclipse.egit.ui.DiffAddBackgroundColor=11,121,90
+org.eclipse.egit.ui.DiffAddForegroundColor=216,254,245
+org.eclipse.egit.ui.DiffHeadlineBackgroundColor=71,71,71
+org.eclipse.egit.ui.DiffHeadlineForegroundColor=242,242,242
+org.eclipse.egit.ui.DiffHunkBackgroundColor=53,97,113
+org.eclipse.egit.ui.DiffHunkForegroundColor=233,242,254
+org.eclipse.egit.ui.DiffRemoveBackgroundColor=117,2,36
+org.eclipse.egit.ui.DiffRemoveForegroundColor=255,232,237
+org.eclipse.egit.ui.IgnoredResourceBackgroundColor=47,47,47
+org.eclipse.egit.ui.IgnoredResourceForegroundColor=120,120,120
+org.eclipse.egit.ui.UncommittedChangeBackgroundColor=47,47,47
+org.eclipse.egit.ui.UncommittedChangeForegroundColor=114,157,186
+org.eclipse.gef.color.button=64,64,64
+org.eclipse.gef.color.line.foreground=64,64,64
+org.eclipse.gef.color.list.background=60,60,60
+org.eclipse.gef.color.list.foreground=224,224,224
+org.eclipse.gef.color.list.hover.background=70,70,70
+org.eclipse.gef.color.list.selected.background=90,90,90
+org.eclipse.gef.color.menu.background=40,40,40
+org.eclipse.gef.color.menu.foreground=224,224,224
+org.eclipse.gef.color.menu.foreground.selected=105,105,105
+org.eclipse.jdt.ui.ColoredLabels.inherited=143,143,191
+org.eclipse.jdt.ui.ColoredLabels.writeaccess_highlight=255,128,128
+org.eclipse.jdt.ui.Javadoc.backgroundColor=52,57,61
+org.eclipse.jdt.ui.Javadoc.foregroundColor=238,238,238
+org.eclipse.jface.REVISION_NEWEST_COLOR=75,44,3
+org.eclipse.jface.REVISION_OLDEST_COLOR=154,113,61
+org.eclipse.mylyn.tasks.ui.colors.category.gradient.end=136,137,133
+org.eclipse.mylyn.tasks.ui.colors.category.gradient.start=85,87,83
+org.eclipse.mylyn.tasks.ui.colors.foreground.past.scheduled=52,101,164
+org.eclipse.mylyn.tasks.ui.colors.foreground.task.active=117,80,123
+org.eclipse.mylyn.tasks.ui.colors.foreground.thisweek.scheduled=85,87,83
+org.eclipse.mylyn.tasks.ui.colors.foreground.today.scheduled=52,101,164
+org.eclipse.pde.genericeditor.extension.editor.color.attribute=196,138,114
+org.eclipse.pde.genericeditor.extension.editor.color.comment=96,139,78
+org.eclipse.pde.genericeditor.extension.editor.color.header=128,128,128
+org.eclipse.pde.genericeditor.extension.editor.color.quote=118,200,244
+org.eclipse.pde.genericeditor.extension.editor.color.tag=86,156,214
+org.eclipse.pde.genericeditor.extension.editor.color.text=212,212,212
+org.eclipse.search.ui.match.highlight=206,92,0
+org.eclipse.ui.editors.rangeIndicatorColor=27,118,153
+org.eclipse.ui.workbench.ACTIVE_NOFOCUS_TAB_BG_END=41,41,41
+org.eclipse.ui.workbench.ACTIVE_NOFOCUS_TAB_BG_START=43,44,45
+org.eclipse.ui.workbench.ACTIVE_NOFOCUS_TAB_TEXT_COLOR=204,204,204
+org.eclipse.ui.workbench.ACTIVE_TAB_BG_END=41,41,41
+org.eclipse.ui.workbench.ACTIVE_TAB_BG_START=43,44,45
+org.eclipse.ui.workbench.ACTIVE_TAB_TEXT_COLOR=221,221,221
+org.eclipse.ui.workbench.FORM_HEADING_ERROR_COLOR=255,110,110
+org.eclipse.ui.workbench.FORM_HEADING_INFO_COLOR=170,170,170
+org.eclipse.ui.workbench.FORM_HEADING_WARNING_COLOR=255,200,0
+org.eclipse.ui.workbench.HOVER_BACKGROUND=52,57,61
+org.eclipse.ui.workbench.HOVER_FOREGROUND=238,238,238
+org.eclipse.ui.workbench.INACTIVE_TAB_BG_END=49,53,56
+org.eclipse.ui.workbench.INACTIVE_TAB_BG_START=59,64,66
+org.eclipse.ui.workbench.INACTIVE_TAB_TEXT_COLOR=187,187,187
+org.eclipse.ui.workbench.INFORMATION_BACKGROUND=81,86,88
+org.eclipse.ui.workbench.INFORMATION_FOREGROUND=238,238,238
+overriddenByCSS=,org.eclipse.egit.ui.CommitMessageCommentColor,org.eclipse.egit.ui.DiffAddBackgroundColor,org.eclipse.egit.ui.DiffAddForegroundColor,org.eclipse.egit.ui.DiffHeadlineBackgroundColor,org.eclipse.egit.ui.DiffHeadlineForegroundColor,org.eclipse.egit.ui.DiffHunkBackgroundColor,org.eclipse.egit.ui.DiffHunkForegroundColor,org.eclipse.egit.ui.DiffRemoveBackgroundColor,org.eclipse.egit.ui.DiffRemoveForegroundColor,org.eclipse.egit.ui.IgnoredResourceBackgroundColor,org.eclipse.egit.ui.IgnoredResourceForegroundColor,org.eclipse.egit.ui.UncommittedChangeBackgroundColor,org.eclipse.egit.ui.UncommittedChangeForegroundColor,org.eclipse.gef.color.line.foreground,org.eclipse.gef.color.list.selected.background,org.eclipse.gef.color.list.hover.background,org.eclipse.gef.color.list.background,org.eclipse.gef.color.list.foreground,org.eclipse.gef.color.menu.background,org.eclipse.gef.color.menu.foreground,org.eclipse.gef.color.menu.foreground.selected,org.eclipse.gef.color.button,org.eclipse.jdt.ui.ColoredLabels.inherited,org.eclipse.jdt.ui.Javadoc.backgroundColor,org.eclipse.jdt.ui.Javadoc.foregroundColor,org.eclipse.jdt.ui.ColoredLabels.writeaccess_highlight,org.eclipse.mylyn.tasks.ui.colors.foreground.today.scheduled,org.eclipse.mylyn.tasks.ui.colors.foreground.past.scheduled,org.eclipse.mylyn.tasks.ui.colors.foreground.thisweek.scheduled,org.eclipse.mylyn.tasks.ui.colors.foreground.task.active,org.eclipse.mylyn.tasks.ui.colors.category.gradient.start,org.eclipse.mylyn.tasks.ui.colors.category.gradient.end,org.eclipse.pde.genericeditor.extension.editor.color.tag,org.eclipse.pde.genericeditor.extension.editor.color.attribute,org.eclipse.pde.genericeditor.extension.editor.color.text,org.eclipse.pde.genericeditor.extension.editor.color.header,org.eclipse.pde.genericeditor.extension.editor.color.comment,org.eclipse.pde.genericeditor.extension.editor.color.quote,ACTIVE_HYPERLINK_COLOR,CONFLICTING_COLOR,CONTENT_ASSIST_BACKGROUND_COLOR,CONTENT_ASSIST_FOREGROUND_COLOR,org.eclipse.ui.workbench.INFORMATION_BACKGROUND,org.eclipse.ui.workbench.INFORMATION_FOREGROUND,org.eclipse.ui.workbench.HOVER_BACKGROUND,org.eclipse.ui.workbench.HOVER_FOREGROUND,org.eclipse.ui.workbench.FORM_HEADING_ERROR_COLOR,org.eclipse.ui.workbench.FORM_HEADING_WARNING_COLOR,org.eclipse.ui.workbench.FORM_HEADING_INFO_COLOR,ERROR_COLOR,HYPERLINK_COLOR,INCOMING_COLOR,OUTGOING_COLOR,RESOLVED_COLOR,EDITION_COLOR,org.eclipse.search.ui.match.highlight,org.eclipse.ui.editors.rangeIndicatorColor,org.eclipse.jface.REVISION_NEWEST_COLOR,org.eclipse.jface.REVISION_OLDEST_COLOR,
