+++ Servers/.settings/org.eclipse.wst.server.core.prefs
@@ -1,0 +1,3 @@
+#Thu Apr 04 09:43:25 KST 2024
+org.eclipse.wst.server.core.isServerProject=true
+eclipse.preferences.version=1
