--- theJoeunUniversity/theJoeunUniversity/.settings/org.eclipse.wst.common.component
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?><project-modules id="moduleCoreId" project-version="1.5.0">
-    <wb-module deploy-name="theJoeunUniversity">
-        <wb-resource deploy-path="/" source-path="/src/main/webapp" tag="defaultRootSource"/>
-        <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/java"/>
-        <property name="context-root" value="theJoeunUniversity"/>
-        <property name="java-output-path" value="/theJoeunUniversity/build/classes"/>
-    </wb-module>
-</project-modules>
