--- theJoeunUniversity/theJoeunUniversity/src/main/webapp/META-INF/MANIFEST.MF
@@ -1,3 +1,0 @@
-Manifest-Version: 1.0
-Class-Path: 
-
